Start to port analyzer.
- Port analyzer and make it accessible from the REPL. Add respective
runtime data primitives to cljc.core. Doesn't work yet, because
sample.metacircular building seems to depend on init_cljc_DOT_*
constructor order in driver.c, either the REGEX constants are not
properly initialized or the vtable of the methods is null (?).

- Port printing to new IPrintWithWriter protocol and move string
  conversion to str. Isolation of IPrintable protocol still TODO.

- Add UUID. Fix reading of sets and other '#' initialized objects.

- Make StringBuilder mutable and behave like in Java or JavaScript, not
returning a new StringBuilder on -append!. This makes code more easily
portable.

- Date type still missing as well as port of macro system removed for
  now.